Dog on the platform
Dog teleports to the platform and juanito teleports to the side.  Dog weighs down the platform and you can press switch to move the platform down.
