+++ Assets/Scripts/Pulleys/pulley_manager.cs.meta
