+++ Assets/Scripts/Pulleys/pulley_switch_manager.cs
