--- Assets/food.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: a9fd52f2b7fb5af47aab320c11499697
-timeCreated: 1525133988
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
